add new page quad rotation

+++ Voidstar/src/Application.h
@@ -12,13 +12,13 @@ namespace Voidstar
 		void Run();
 
 		std::unique_ptr<Camera>& GetCamera() { return m_Camera; }
-		size_t GetExeTime() { return m_ExeTime; }
+		float GetExeTime() { return m_ExeTime; }
 		virtual void Execute() {};
 		virtual ~Application();
 	private:
 		friend class Renderer;
 		size_t m_ScreenWidth, m_ScreenHeight;
-		size_t m_ExeTime = 0;
+		float m_ExeTime = 0;
 		UPtr<Camera> m_Camera;
 		SPtr<Window> m_Window;
 	};
